requirements.txt
# local package
-e .

# external requirements
pyTelegramBotAPI
python-dotenv
requests
json
pandas
yfinance
DateTime
matplotlib
scikit-learn
numpy
mysql-connector-python
